Explain (at length) the use of link text vs target
Link text is what appears on the page.  Target is the page the link
points to.  We use link text as both the presented name and the unique ID
of each actor and actress.  I began implementing a switch to link target,
and I reverted it.

This comment captures my reasons.