Linkify all names identically, per film
If the same name appears twice (e.g., stars list and cast section, or
director and producer) and it's a link in only one of those places,
convert the plain-text places to also be a link.

Do this on a per-film basis so we don't accidentally promote an obscure
John Doe unworthy of a Wikipedia page to merge with a more popular John
Doe who warrants a link in some other movie.  The corner case here is a
single movie with two actors -- one obscure, one not -- who have the same
name.  That's tough to fix without NLP.