Strip epithets, block directors
"The Cannonball Run" has epithets before some cast members names:
  * Director [[Hal Needham]] appears uncredited as the ambulance EMT.
  * Veteran Daytona 500 commentator [[Ken Squier]] appears as a ...
  * Veteran voice actor [[June Foray]] provided the dubbed dialogue of ...
Strip out those epithets so we emit Hal Needham, Ken Squier, and June
Foray as actors.

"Run for Love" lists directors among the cast, even though they are not
cast:
  *Director: [[Zhang Yibai]]

This is hard to distinguish from the form in The Cannonball Run!  This
commit uses /^Director\s+/ as an epithet meaning that the director played
a role, and /^Director\S+/ as indicating who directed the film.  The
former generates an actor entry, and the latter does not.