Parse `Name` column in cast table
When the `==Cast==` section is a big table, we look for columns with a
header like `Actor` or `Actress` that indicates that the cells in that
column are the names of the performers in the film.  This commit adds
`Name` as an indicator of such a column, but only if there isn't already
an `Actor`, `Actress`, or `Cast` column.  (If such a column exists, then
`Name` might be a character name.)